Update: middleware del & socket connection live
Deleted middleware from bolt documentation and
solved the problem of the app not recieving
events. The problem was that Event Subscriptions
list was blank so my bot was not listening for events
within the scope I had allowed it from Bot Token
Scopes. So my bot had the scope but was not listening
for what I wanted this whole time;